app/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0

# Database
sqlalchemy==2.0.44
sqlmodel==0.0.16
psycopg2-binary==2.9.11
alembic==1.17.2

# ORM + Pydantic
pydantic==2.12.4
pydantic-core==2.41.5

# Security / Auth
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# File upload + static
python-multipart==0.0.20
aiofiles==25.1.0

# Email
fastapi-mail==1.4.1

# Utils
python-dotenv==1.2.1
pdfplumber==0.10.3
python-docx==1.1.0
openai==1.12.0
typing-extensions==4.15.0
httpx==0.28.1
httpcore==1.0.9

# PDF + Image
pypdfium2==5.0.0
Pillow==12.0.0

# SQLAdmin (YOU USE IT — NOT IN YOUR REQUIREMENTS)
sqladmin==0.20.0

# Web dependencies
starlette==0.49.3
anyio==4.11.0
passlib==1.7.4